rename : category 도메인 내에 exception 패키지 위치 수정
